# recruitment-automation-system-ui/src/app/services/models/application-request.ts
/* tslint:disable */
/* eslint-disable */
export interface ApplicationRequest {
  applicationDate?: string;
  candidateEmail: string;
  candidateName: string;
  coverLetter?: string;
  id?: number;
  jobId: number;
  resumeUrl: string;
  status?: string;
}