# recruitment-automation-system-ui/src/app/services/models/application-response.ts
/* tslint:disable */
/* eslint-disable */
export interface ApplicationResponse {
  applicationDate?: string;
  candidateEmail?: string;
  candidateName?: string;
  coverLetter?: string;
  id?: number;
  resumeUrl?: string;
  status?: string;
}